Add sad path tests for missing field and email taken

+++ spec/requests/api/v1/weather_request_spec.rb
@@ -58,7 +58,7 @@ describe 'Weather API' do
       expect(daily_weather).to have_key(:sunset)
       expect(daily_weather[:sunset]).to be_a String
       expect(daily_weather).to have_key(:max_temp)
-      expect(daily_weather[:max_temp]).to be_a Float
+      expect(daily_weather[:max_temp]).to be_a Numeric
       expect(daily_weather).to have_key(:min_temp)
       expect(daily_weather[:min_temp]).to be_a Numeric
       expect(daily_weather).to have_key(:conditions)
